+++ docs/changelog.md
@@ -17,6 +17,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: add developer documentation for service classes
 * chore: format dialog action class
 * feat: project specific subtask templates
+* refactor: remove dependency on user from subtask template service
 
 ### [22.11.1] - 2022-11-21